backend/requirements.txt
django==2.1
djangorestframework==3.8
gunicorn
markdown
pygments
requests
pillow
faker
factory_boy

